Make bMap be an AVL tree, change the way some funcs work
When we add a node to bm (a bMap *), we replace bm with the results of the
function addNode

We do the same whenever we perform a right or a left rotation. Right rotations
move the left child to where the operand is. Left rotation does the opposite,
replacing the right child of the operand with the operand.